Create individual project for Data, Model and Controller
