+++ Bulky.DataAccess/Repository/IRepository/IRepository.cs
@@ -1,0 +1,8 @@
+using System;
+
+namespace Bulky.DataAccess.Repository.IRepository;
+
+public interface IRepository
+{
+
+}
