+++ Bulky.Utility/SD.cs
@@ -1,0 +1,8 @@
+using System;
+
+namespace Bulky.Utility;
+
+public static class SD
+{
+
+}
