+++ BulkyWeb/Controllers/CategoryController.cs
@@ -1,6 +1,6 @@
 using System.Data.Common;
-using BulkyWeb.Data;
-using BulkyWeb.Models;
+using Bulky.DataAccess.Data;
+using Bulky.Models;
 using Microsoft.AspNetCore.Mvc;
 
 namespace BulkyWeb.Controllers
