+++ BulkyWeb/Controllers/HomeController.cs
@@ -1,17 +1,12 @@
 using System.Diagnostics;
+using Bulky.Models;
 using Microsoft.AspNetCore.Mvc;
-using BulkyWeb.Models;
 
 namespace BulkyWeb.Controllers;
 
-public class HomeController : Controller
+public class HomeController(ILogger<HomeController> logger) : Controller
 {
-    private readonly ILogger<HomeController> _logger;
-
-    public HomeController(ILogger<HomeController> logger)
-    {
-        _logger = logger;
-    }
+    private readonly ILogger<HomeController> _logger = logger;
 
     public IActionResult Index()
     {
