+++ BulkyWeb/Views/Category/Delete.cshtml
@@ -1,34 +1,41 @@
 ﻿@model Category
 
-<form method="POST">
-    <input asp-for="Id" hidden/>
-    <div class="border p-3 mt-4">
-        <div class="row pb-2">
-            <h2 class="text-primary">Delete Category</h2>
-            <hr />
-        </div>
-        
-        <div class="mb-3 row p-1">
-            <label asp-for="Name" class="p-0"></label>
-            <input asp-for="Name" class="form-control" disabled/>
-        </div>
-
-        <div class="mb-3 row p-1">
-            <label asp-for="DisplayOrder" class="p-0"></label>
-            <input asp-for="DisplayOrder" class="form-control" disabled/>
-        </div>
-        <div class="row py-2">
-            <div class="col-6 col-md-3">
-                <button type="submit" class="btn btn-danger form-control">Delete</button>
-            </div>
-            <div class="col-6 col-md-3">
-                <a asp-controller="Category" asp-action="Index" class="btn btn-secondary border form-control">
-                    Back to List
-                </a>
+<div class="card shadow border-0 mt-4">
+    <div class="card-header bg-secondary bg-gradient ml-0 py-3">
+        <div class="row">
+            <div class="col-12 text-center">
+                <h2 class="text-white py-2">Delete Category</h2>
             </div>
         </div>
     </div>
-</form>
+    <div class="card-body py-4">
+        <form method="POST">
+            <input asp-for="Id" hidden />
+            <div class="border p-3">
+                <div class="form-floating py-2 col-12">
+                    <input asp-for="Name" class="form-control border-0 shadow" disabled/>
+                    <label asp-for="Name" class="ms-2"></label>
+                </div>
+
+                <div class="form-floating py-2 col-12">
+                    <input asp-for="DisplayOrder" class="form-control border-0 shadow" disabled/>
+                    <label asp-for="DisplayOrder" class="ms-2"></label>
+                </div>
+                <div class="row pt-2">
+                    <div class="col-6 col-md-3">
+                        <button type="submit" class="btn btn-danger form-control">Delete</button>
+                    </div>
+                    <div class="col-6 col-md-3">
+                        <a asp-controller="Category" asp-action="Index" class="btn btn-secondary border form-control">
+                            Back to List
+                        </a>
+                    </div>
+                </div>
+            </div>
+        </form>
+    </div>
+</div>
+
 
 @section Scripts {
     @{
