+++ BulkyWeb/Views/Category/Index.cshtml
@@ -1,22 +1,24 @@
 ﻿@model List<Category>
 
-<div class="container">
-    <div class="row pt-4 pb-4">
-        <div class="col-6">
-            <h2 class="text-primary">
-                Category List
-            </h2>
+<div class="card shadow border-0 mt-4">
+    <div class="card-header bg-secondary bg-gradient ml-0 py-3">
+        <div class="row">
+            <div class="col-12 text-center">
+                <h2 class="text-white py-2">Category List</h2>
+            </div>
         </div>
-
-        <div class="col-6 text-end">
-            <a asp-controller="Category" asp-action="Create" class="btn btn-primary">
-                <i class="bi bi-plus-circle"></i> Create New Category
-            </a>
-        </div>
-        
     </div>
-</div>
-<table class="table table-bordered table-striped">
+    
+    <div class="card-body py-4">
+        <div class="row pt-4 pb-4">
+            <div class="col-6"></div>
+            <div class="col-6 text-end">
+                <a asp-controller="Category" asp-action="Create" class="btn btn-primary">
+                    <i class="bi bi-plus-circle"></i> Create New Category
+                </a>
+            </div>
+        </div>
+        <table class="table table-bordered table-striped">
     <thead>
         <th>
             Category Name
@@ -45,3 +47,9 @@
         }
     </tbody>
 </table>
+    </div>
+</div>
+
+<div class="container">
+    
+</div>
