+++ BulkyWeb/Views/_ViewImports.cshtml
@@ -1,3 +1,3 @@
 ﻿@using BulkyWeb
-@using BulkyWeb.Models
+@using Bulky.Models
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
